modernize AdminPeopleApp [skip ci]
